Completed permission control in views and base.html file

+++ main/views.py
@@ -6,8 +6,13 @@ from django.conf import settings
 from django.contrib.auth.models import User
 
 
+def no_access(request):
+    return render(request, 'no_access.html', {})
+
 def home(request):
     context = {}
+    if not request.user.is_authenticated():
+        return redirect("github_login")
     user = request.user
     context['user'] = user
     if request.method=="POST":
@@ -31,6 +36,8 @@ def home(request):
 
 def create_coffee(request):
     context = {}
+    if not request.user.is_authenticated():
+        return redirect("github_login")
     if request.method == "POST":
         form = CoffeeForm(request.POST)
         context['form'] = form
@@ -50,8 +57,15 @@ def create_coffee(request):
 
 def edit_coffee(request, coffee_id):
     context = {}
+    if not request.user.is_authenticated():
+        return redirect("github_login")
+
     coff = Coffee.objects.get(id=coffee_id)
     context['coffee'] = coff
+
+    if not (request.user == coff.user or request.user.is_superuser or request.user.is_staff):
+        return redirect("no_access")
+
     if request.method == "POST":
         form = CoffeeForm(request.POST, instance=coff)
         context['form'] = form
@@ -70,11 +84,18 @@ def edit_coffee(request, coffee_id):
 
 
 def delete_coffee(request, coffee_id):
-    Coffee.objects.get(id=coffee_id).delete()
+    if not request.user.is_authenticated():
+        return redirect("github_login")
+    coffee = Coffee.objects.get(id=coffee_id)
+    if not (request.user == coffee.user or request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
+    coffee.delete()
     return redirect("home")
 
 def create_bean(request):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     if request.method == "POST":
         form = BeanForm(request.POST)
         context['form'] = form
@@ -91,6 +112,8 @@ def create_bean(request):
 
 def edit_bean(request, bean_id):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     bean = Bean.objects.get(id=bean_id)
     context['bean']=bean
     if request.method == "POST":
@@ -107,11 +130,15 @@ def edit_bean(request, bean_id):
         return render(request, 'edit_bean.html', context)
 
 def delete_bean(request, bean_id):
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     Bean.objects.get(id=bean_id).delete()
     return redirect("home")
 
 def create_roast(request):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     if request.method == "POST":
         form = RoastForm(request.POST)
         context['form'] = form
@@ -128,6 +155,8 @@ def create_roast(request):
 
 def edit_roast(request, roast_id):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     roast = Roast.objects.get(id=roast_id)
     context['roast']=roast
     if request.method == "POST":
@@ -144,11 +173,15 @@ def edit_roast(request, roast_id):
         return render(request, 'edit_roast.html', context)
 
 def delete_roast(request, roast_id):
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     Roast.objects.get(id=roast_id).delete()
     return redirect("home")
 
 def create_powder(request):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     if request.method == "POST":
         form = PowderForm(request.POST)
         context['form'] = form
@@ -165,6 +198,8 @@ def create_powder(request):
 
 def edit_powder(request, powder_id):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     powder = Powder.objects.get(id=powder_id)
     context['powder']=powder
     if request.method == "POST":
@@ -181,11 +216,15 @@ def edit_powder(request, powder_id):
         return render(request, 'edit_powder.html', context)
 
 def delete_powder(request, powder_id):
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     Powder.objects.get(id=powder_id).delete()
     return redirect("home")
 
 def create_syrup(request):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     if request.method == "POST":
         form = SyrupForm(request.POST)
         context['form'] = form
@@ -202,6 +241,8 @@ def create_syrup(request):
 
 def edit_syrup(request, syrup_id):
     context = {}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     syrup = Syrup.objects.get(id=syrup_id)
     context['syrup']=syrup
     if request.method == "POST":
@@ -218,11 +259,15 @@ def edit_syrup(request, syrup_id):
         return render(request, 'edit_syrup.html', context)
 
 def delete_syrup(request, syrup_id):
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     Syrup.objects.get(id=syrup_id).delete()
     return redirect("home")
 
 def create_order(request, coffee_id):
     context = {}
+    if not request.user.is_authenticated():
+        return redirect("github_login")
     coffee = Coffee.objects.get(id=coffee_id)
     context['coffee'] = coffee
     if request.method == "POST":
@@ -243,6 +288,8 @@ def create_order(request, coffee_id):
 
 def place_order(request, year, month, day):
     context = {}
+    if not request.user.is_authenticated():
+        return redirect("github_login")
     date = datetime.datetime.strptime('%s%s%s'%(year, month, day), "%Y%m%d").date()
     order_list=Order.objects.filter(user=request.user, date=date)
     subject = "My Coffee Orders:"
@@ -254,6 +301,8 @@ def place_order(request, year, month, day):
 
 def replicate_order(request, year, month, day):
     context = {}
+    if not request.user.is_authenticated():
+        return redirect("github_login")
     date = datetime.datetime.strptime('%s%s%s'%(year, month, day), "%Y%m%d").date()
     context['date']=date
     if request.method=="POST":
@@ -272,12 +321,16 @@ def replicate_order(request, year, month, day):
 
 def user_list(request):
     context={}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     users = User.objects.all()
     context['users']=users
     return render(request, 'user_list.html', context)
 
 def user_coffees(request, user_id):
     context={}
+    if not (request.user.is_superuser or request.user.is_staff):
+            return redirect("no_access")
     user = User.objects.get(id=user_id)
     coffee_list = Coffee.objects.filter(user=user)
     context['user']=user
